% Auto-generated: do not edit by hand
\name{''StreamReaderList}

\alias{''StreamReaderList}

\title{StreamReaderList component}

\description{

}

\usage{
''StreamReaderList(id=NULL, style=NULL, url=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{style}{Named list. The style used on the container.}

\item{url}{Character. The URL used to fetcch data.}

\item{value}{Numeric. The value used on the counter.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

